salvage real unit tests: - move actual unit tests from the old (mostly integration-test-heavy) platform.spec to new spec-cordova/platform.spec.js.
- moved the actual unit tests for cordova.platform out of the integration tests folder and back into the spec folder.
